buildr confirms project 2 works
would be nice to add asserts... but that's not my goal right now.
symbolic logic class awaits!
