+++ assignments/src/pset/two/ReadWriteLock/Writer.java
@@ -1,19 +1,21 @@
+package pset.two.ReadWriteLock;
+
 /*
  * Decompiled with CFR 0_110.
  * 
  * Could not load the following classes:
  *  ReadWriteLock
  */
-final class Reader
+public final class Writer
 extends ReaderWriter {
-    public Reader(ReadWriteLock readWriteLock, StringBuffer stringBuffer, int n) {
+    public Writer(ReadWriteLock readWriteLock, StringBuffer stringBuffer, int n) {
         super(readWriteLock, stringBuffer, n);
     }
 
     @Override
-    public void beginSession() {
+    protected void beginSession() {
         try {
-            this.lock.beginRead();
+            this.lock.beginWrite();
         }
         catch (Exception var1_1) {
             var1_1.printStackTrace();
@@ -22,13 +24,13 @@ extends ReaderWriter {
 
     @Override
     protected void inSession() {
-        this.sb.append("Reader reads " + this.id + "\n");
+        this.sb.append("Writer writes " + this.id + "\n");
     }
 
     @Override
     protected void endSession() {
         try {
-            this.lock.endRead();
+            this.lock.endWrite();
         }
         catch (Exception var1_1) {
             var1_1.printStackTrace();
