+++ assignments/test/pset/two/CyclicBarrier/CyclicBarrierTest.java
@@ -1,3 +1,5 @@
+package pset.two.CyclicBarrier;
+
 /*
  * Decompiled with CFR 0_110.
  * 
@@ -9,11 +11,17 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 
+import static org.junit.Assert.*;
+import org.junit.Test;
+
+import pset.two.CyclicBarrier.CyclicBarrier;
+
 public class CyclicBarrierTest {
     static final int SIZE = 10;
     static final int ROUND = 2;
 
-    public static void main(String[] arrstring) {
+    @Test
+    public void testCyclicBarrier() {
         CyclicBarrier cyclicBarrier = new CyclicBarrier(10);
         ExecutorService executorService = Executors.newCachedThreadPool();
         for (int i = 0; i < 10; ++i) {
@@ -22,8 +30,7 @@ public class CyclicBarrierTest {
         executorService.shutdown();
     }
 
-    public static class TestThread
-    implements Runnable {
+    public static class TestThread implements Runnable {
         final CyclicBarrier barrier_;
 
         public TestThread(CyclicBarrier cyclicBarrier) {
@@ -37,13 +44,11 @@ public class CyclicBarrierTest {
                 System.out.println("Thread " + Thread.currentThread().getId() + " is WAITING round:" + i);
                 try {
                     n = this.barrier_.await();
-                }
-                catch (Exception var3_3) {
+                } catch (Exception var3_3) {
                     var3_3.printStackTrace();
                 }
                 System.out.println("Thread " + Thread.currentThread().getId() + " got index:" + n);
             }
         }
     }
-
 }
